Refactor: Extract SSDPSocket from SonosController
Phase 1 of god object refactoring: Extract low-level networking infrastructure.

- Create Infrastructure/SSDPSocket.swift (80 lines)
- Extract BSD socket implementation from SonosController
- Fix deprecated String(cString:) usage with proper UTF-8 decoding
- Reduce SonosController from 1,732 to 1,654 lines (-78)

This is a pure extraction with no behavior changes. The Socket class is now:
- Properly scoped to SSDP discovery
- Better documented with inline comments
- Follows modern Swift best practices
- Can be easily unit tested in isolation

Related to P1 Architecture task in ROADMAP.md:61
