+++ SonosVolumeController/Sources/SonosController.swift
@@ -149,7 +149,7 @@ actor SonosController {
         """
 
         do {
-            let socket = try Socket()
+            let socket = try SSDPSocket()
 
             // Send multiple discovery packets to catch devices that might miss the first one
             print("📡 Sending discovery packet 1/3...")
@@ -1651,82 +1651,3 @@ actor SonosController {
     }
 }
 
-// BSD socket for SSDP discovery
-import Darwin
-
-class Socket {
-    private var sockfd: Int32 = -1
-
-    init() throws {
-        sockfd = socket(AF_INET, SOCK_DGRAM, 0)
-        guard sockfd >= 0 else {
-            throw NSError(domain: "Socket", code: Int(errno), userInfo: [NSLocalizedDescriptionKey: "Failed to create socket"])
-        }
-
-        // Allow socket reuse
-        var reuseAddr: Int32 = 1
-        setsockopt(sockfd, SOL_SOCKET, SO_REUSEADDR, &reuseAddr, socklen_t(MemoryLayout<Int32>.size))
-
-        #if os(macOS)
-        var reusePort: Int32 = 1
-        setsockopt(sockfd, SOL_SOCKET, SO_REUSEPORT, &reusePort, socklen_t(MemoryLayout<Int32>.size))
-        #endif
-
-        // Set timeout for receives
-        var timeout = timeval(tv_sec: 1, tv_usec: 0)
-        setsockopt(sockfd, SOL_SOCKET, SO_RCVTIMEO, &timeout, socklen_t(MemoryLayout<timeval>.size))
-    }
-
-    deinit {
-        if sockfd >= 0 {
-            close(sockfd)
-        }
-    }
-
-    func send(_ message: String, to address: String, port: UInt16) throws {
-        guard let data = message.data(using: .utf8) else { return }
-
-        var addr = sockaddr_in()
-        addr.sin_family = sa_family_t(AF_INET)
-        addr.sin_port = port.bigEndian
-        inet_pton(AF_INET, address, &addr.sin_addr)
-
-        let sent = data.withUnsafeBytes { (buffer: UnsafeRawBufferPointer) -> Int in
-            return withUnsafePointer(to: addr) { addrPtr in
-                addrPtr.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockaddrPtr in
-                    return sendto(sockfd, buffer.baseAddress, buffer.count, 0, sockaddrPtr, socklen_t(MemoryLayout<sockaddr_in>.size))
-                }
-            }
-        }
-
-        guard sent >= 0 else {
-            throw NSError(domain: "Socket", code: Int(errno), userInfo: [NSLocalizedDescriptionKey: "Failed to send"])
-        }
-    }
-
-    func receive(timeout: TimeInterval) throws -> (data: String, address: String)? {
-        var buffer = [UInt8](repeating: 0, count: 8192)
-        var addr = sockaddr_in()
-        var addrLen = socklen_t(MemoryLayout<sockaddr_in>.size)
-
-        let received = withUnsafeMutablePointer(to: &addr) { addrPtr in
-            addrPtr.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockaddrPtr in
-                recvfrom(sockfd, &buffer, buffer.count, 0, sockaddrPtr, &addrLen)
-            }
-        }
-
-        guard received > 0 else {
-            return nil
-        }
-
-        guard let string = String(bytes: buffer[..<received], encoding: .utf8) else {
-            return nil
-        }
-
-        var hostBuffer = [CChar](repeating: 0, count: Int(INET_ADDRSTRLEN))
-        inet_ntop(AF_INET, &addr.sin_addr, &hostBuffer, socklen_t(INET_ADDRSTRLEN))
-        let host = String(cString: hostBuffer)
-
-        return (string, host)
-    }
-}
